TagList Redux refactor
